# openapi.json
{
  "openapi": "3.1.0",
  "components": {
    "securitySchemes": {
      "x402": {
        "type": "apiKey",
        "in": "header",
        "name": "Payment-Signature",
        "description": "x402 v2 payment payload (base64)"
      }
    }
  },
  "info": {
    "title": "x402 Trust Layer",
    "version": "5.2.1",
    "description": "68 paid x402 trust infrastructure APIs: guard, semantic escrow, mandate diff, certified seller network, KYM, mandates, compliance, disputes, and orchestration.",
...
  },
...
  "servers": [
    {
      "url": "https://x402trustlayer.xyz"
    }
  ],
  "paths": {
    "/api/a2a/execute": {
      "post": {
        "operationId": "post_a2a_execute",
        "summary": "Agent-to-agent x402 orchestration: trust preflight then paid call to seller endpoint",
        "description": "Agent-to-agent x402 orchestration: trust preflight then paid call to seller endpoint \u2014 $0.10 USDC via x402 (Dexter facilitator https://x402.dexter.cash).",
        "tags": [
          "killer",
          "a2a",
          "agents",
          "orchestration"
        ],
        "security": [
          {
            "x402": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response after x402 settlement",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Successful response after x402 settlement",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "allowed": {
                      "type": "boolean"
                    },
                    "summary": {
                      "type": "string"
                    },
                    "confidence": {
                      "type": "number"
                    },
                    "checks_passed": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "additionalProperties": true
                },
                "example": {
                  "success": true,
                  "allowed": true,
                  "orchestratorReady": true,
                  "orchestration": "in-process",
                  "sellerResponse": {
                    "ok": true,
                    "allowed": true,
                    "summary": "Guard preflight passed for A2A seller call"
                  },
                  "checks_passed": [
                    "a2a_schema_valid",
                    "trust_score",
                    "seller_dispatch"
                  ]
                }
              }
            }
          },
          "402": {
            "description": "Payment Required (x402 v2). Unpaid requests must receive HTTP 402 with Payment-Required header and non-empty accepts.",
            "headers": {
              "Payment-Required": {
                "description": "Base64-encoded x402 payment requirements (v2)",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.100000"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "buyerAgentId": {
                    "type": "string"
                  },
                  "sellerAgentId": {
                    "type": "string"
                  },
                  "sellerEndpoint": {
                    "type": "string"
                  },
                  "taskDescription": {
                    "type": "string"
                  },
                  "maxBudgetUsdc": {
                    "type": "number"
                  }
                },
                "required": [
                  "buyerAgentId",
                  "sellerAgentId",
                  "sellerEndpoint",
                  "taskDescription",
                  "maxBudgetUsdc"
                ]
              },
              "example": {
                "buyerAgentId": "dexter-verifier-probe",
                "sellerAgentId": "seller-oracle-1",
                "sellerEndpoint": "https://x402trustlayer.xyz/api/guard/pre-x402",
                "taskDescription": "Preflight guard check for ETH oracle call under $0.10",
                "maxBudgetUsdc": 0.1
              }
            }
          }
        }
      }
    },
...
    "/api/agent/verify": {
      "post": {
        "operationId": "post_agent_verify",
        "summary": "ERC-8004 TrustScore on Base mainnet",
        "description": "ERC-8004 TrustScore on Base mainnet \u2014 $0.04 USDC via x402 (Dexter facilitator https://x402.dexter.cash).",
        "tags": [
          "identity",
          "erc-8004",
          "identity",
          "trust-score"
        ],
        "security": [
          {
            "x402": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response after x402 settlement",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Successful response after x402 settlement",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "allowed": {
                      "type": "boolean"
                    },
                    "summary": {
                      "type": "string"
                    },
                    "confidence": {
                      "type": "number"
                    },
                    "checks_passed": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "additionalProperties": true
                },
                "example": {
                  "ok": true,
                  "allowed": true,
                  "summary": "Paid response after x402 settlement"
                }
              }
            }
          },
          "402": {
            "description": "Payment Required (x402 v2). Unpaid requests must receive HTTP 402 with Payment-Required header and non-empty accepts.",
            "headers": {
              "Payment-Required": {
                "description": "Base64-encoded x402 payment requirements (v2)",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.040000"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "walletAddress": {
                    "type": "string"
                  },
                  "agentId": {
                    "type": "string"
                  }
                },
                "required": [
                  "walletAddress",
                  "agentId"
                ]
              },
              "example": {
                "walletAddress": "0x0000000000000000000000000000000000000001",
                "agentId": "1"
              }
            }
          }
        }
      }
    },
...
  }
}